Refactored into new class
